#159. Extracted utils to pkg. Created internal pkg doctor-schema and logging tas these methods do not need to be exposed as of now.
