+++ src/pkg/utils/files.go
@@ -23,19 +23,6 @@ func WriteOutput(output []byte, outputFile string) (err error) {
 	return nil
 }
 
-// OpenLogFile opens the provided input file or returns an error if the file does not exist.
-func OpenLogFile(path string) (file *os.File, err error) {
-	err = CreateFileDirs(path)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create log file path: %s\n", err)
-	}
-	file, err = os.OpenFile(path, os.O_CREATE|os.O_WRONLY|os.O_APPEND, 0644)
-	if err != nil {
-		return nil, fmt.Errorf("failed to open log file: %s\n", err)
-	}
-	return file, nil
-}
-
 // CreateFileDirs creates the directories for a file path if they do not exist.
 func CreateFileDirs(fullPath string) (err error) {
 	pathSeperator := string(os.PathSeparator)
