+++ src/pkg/utils/files_test.go
@@ -4,7 +4,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/defenseunicorns/go-oscal/src/internal/utils"
+	"github.com/defenseunicorns/go-oscal/src/pkg/utils"
 )
 
 func TestFileUtils(t *testing.T) {
@@ -55,45 +55,6 @@ func TestFileUtils(t *testing.T) {
 			err := utils.WriteOutput([]byte(testOutput), "")
 			if err == nil {
 				t.Error("expected an error, got", err)
-			}
-		})
-	})
-
-	t.Run("OpenLogFile", func(t *testing.T) {
-		t.Parallel()
-		subdir := tmpDir + "openLogFile"
-		testFile := "/test.log"
-
-		t.Run("It opens the provided log file", func(t *testing.T) {
-			t.Parallel()
-			file, err := utils.OpenLogFile(tmpDir + testFile)
-			if err != nil {
-				t.Error("expected no error, got", err)
-			}
-			if file == nil {
-				t.Error("expected a file, got", file)
-			}
-		})
-
-		t.Run("creates the file path if it doesn't exist", func(t *testing.T) {
-			t.Parallel()
-			file, err := utils.OpenLogFile(subdir + testFile)
-			if err != nil {
-				t.Error("expected no error, got", err)
-			}
-			if file == nil {
-				t.Error("expected a file, got", file)
-			}
-		})
-
-		t.Run("returns an error if no log file is provided", func(t *testing.T) {
-			t.Parallel()
-			file, err := utils.OpenLogFile("")
-			if err == nil {
-				t.Error("expected an error, got", err)
-			}
-			if file != nil {
-				t.Error("expected no file, got", file)
 			}
 		})
 	})
