+++ src/pkg/utils/version_test.go
@@ -6,7 +6,7 @@ import (
 
 	"time"
 
-	"github.com/defenseunicorns/go-oscal/src/internal/utils"
+	"github.com/defenseunicorns/go-oscal/src/pkg/utils"
 	"gopkg.in/yaml.v3"
 )
