+++ src/pkg/validation/validationError.go
@@ -3,7 +3,7 @@ package validation
 import (
 	"strings"
 
-	"github.com/defenseunicorns/go-oscal/src/internal/utils"
+	"github.com/defenseunicorns/go-oscal/src/pkg/utils"
 	"github.com/santhosh-tekuri/jsonschema/v5"
 )
