+++ src/pkg/validation/validator_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/defenseunicorns/go-oscal/src/gooscaltest"
-	"github.com/defenseunicorns/go-oscal/src/internal/utils"
+	"github.com/defenseunicorns/go-oscal/src/pkg/utils"
 	"gopkg.in/yaml.v3"
 )
